preprocessing single script
